# web/openapi.yaml
openapi: 3.1.0
info:
  title: Powered By Arkos.js
  version: 1.0.0
  description: This API was automatically generated by Arkos.js, read more about
    at www.arkosjs.com
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
...
  /api/triages/{id}/approve:
    post:
      tags:
        - Triages
      summary: Approve and refer triage
      description: Approves a triage and sets referral metadata for patient scheduling
      operationId: approveTriage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - specialtyId
                - consultationTypeId
                - priority
              properties:
                specialtyId:
                  type: string
                consultationTypeId:
                  type: string
                priority:
                  type: string
                  enum:
                    - BAIXA
                    - MEDIA
                    - ALTA
      responses:
        "200":
          description: triage approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageResponse'
        "400":
          description: Invalid input or triage state
        "401":
          description: Authentication required
        "403":
          description: Insufficient permissions
        "404":
          description: triage not found
  /api/triages/{id}/reject:
    post:
      tags:
        - Triages
      summary: Reject triage
      description: Rejects a pending or in-review triage
      operationId: rejectTriage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rejectionReason:
                  type: string
      responses:
        "200":
          description: triage rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriageResponse'
        "400":
          description: Invalid triage state
        "401":
          description: Authentication required
        "403":
          description: Insufficient permissions
        "404":
          description: triage not found
...
components:
  schemas:
...
    TriageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            complaint:
              type: string
            symptomDuration:
              type: string
            symptom:
              type: string
            actionTaken:
              type: string
            reactionAfterAction:
              type: string
            status:
              type: string
              enum:
                - PENDENTE
                - EM_ANALISE
                - CANCELADO
                - REENCAMINHADO
            patientId:
              type: string
            analyzedById:
              type: string
            specialtyId:
              type: string
            consultationTypeId:
              type: string
            priority:
              type: string
              enum:
                - BAIXA
                - MEDIA
                - ALTA
            rejectionReason:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            patient:
              type: object
              properties:
                id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
            analyzedBy:
              type: object
              properties:
                id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
            specialty:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            consultationType:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security: []
...